requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
openai==1.3.0
Pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
flask-wtf==1.2.1
gunicorn==21.2.0
chart.js==4.4.1 